# docs/swagger.yaml
definitions:
...
  models.Doorlock:
    properties:
      description:
        type: string
      doorSerialId:
        type: string
      gatewayId:
        type: string
      id:
        type: integer
      lastOpenTime:
        type: integer
      location:
        type: string
      schedulers:
        items:
          $ref: "#/definitions/models.Scheduler"
        type: array
    type: object
  models.DoorlockCmd:
    properties:
      doorSerialId:
        type: string
      gatewayId:
        type: string
      state:
        type: string
    type: object
  models.DoorlockDelete:
    properties:
      doorSerialId:
        type: string
      gatewayId:
        type: string
    required:
      - doorSerialId
    type: object
...
  models.Scheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorSerialId:
        type: string
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      weekDay:
        type: integer
    type: object
...
  models.SwagCreateDoorlock:
    properties:
      areaId:
        type: integer
      description:
        type: string
      gatewayId:
        type: integer
      location:
        type: string
      schedulerId:
        type: integer
    type: object
...
  models.SwagUpdateDoorlock:
    properties:
      areaId:
        type: integer
      description:
        type: string
      gatewayId:
        type: integer
      id:
        type: integer
      location:
        type: string
      schedulerId:
        type: integer
    type: object
...
  utils.ErrorResponse:
    properties:
      errorMsg:
        type: string
      msg:
        type: string
      statusCode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /v1/doorlock:
    delete:
      consumes:
        - application/json
      description: Delete doorlock using "id" field
      parameters:
        - description: Doorlock Delete payload
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.DoorlockDelete"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Delete Doorlock By ID
    patch:
      consumes:
        - application/json
      description: Update doorlock, must have "id" field
      parameters:
        - description: Fields need to update a doorlock
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.SwagUpdateDoorlock"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Update Doorlock By ID
    post:
      consumes:
        - application/json
      description: Create doorlock
      parameters:
        - description: Fields need to create a doorlock
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.SwagCreateDoorlock"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Doorlock"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Create Doorlock
  /v1/doorlock/{id}:
    get:
      description: find doorlock info by doorlock id
      parameters:
        - description: Doorlock ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Doorlock"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Find Doorlock By ID
  /v1/doorlock/state:
    patch:
      consumes:
        - application/json
      description: Update doorlock state, must have "id" field
      parameters:
        - description: Fields need to update a doorlock state
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.DoorlockCmd"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Update Doorlock State By ID
  /v1/doorlocks:
    get:
      description: find all doorlocks info
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: "#/definitions/models.Doorlock"
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Find All Doorlock
...
swagger: "2.0"